requirements.txt

azure-eventhub
azure-identity
azure-keyvault-secrets
requests
pytest==7.4.2
pytest-mock==3.11.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0